web_control/server/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.29.0
python-socketio==5.13.0
python-multipart==0.0.9
aiofiles==24.1.0
numpy>=1.20.0
opencv-python>=4.5.0
